🛠 Message importation to Q2.550 fix

+++ twake/backend/node/src/services/messages/services/engine/processors/channel-view/index.ts
@@ -70,7 +70,7 @@ export class ChannelViewProcessor {
               ...pkPrefix,
               message_id: reversed.message_id,
             });
-            if (existingThreadRef.thread_id === message.resource.thread_id) {
+            if (existingThreadRef && existingThreadRef.thread_id === message.resource.thread_id) {
               reversed.message_id = message.resource.id;
               await this.repositoryReversed.save(reversed);
               await this.repository.remove(existingThreadRef);
